When calculating the time entries, the restriction on parent issues should NOT be present.
This is easily explainable by looking at how time is logged: in this case, we are not considering the amount
of time spent on a task, but the single time entries. As such, all the time entries are actually logged on
issues which have a parent, but NOT on the parent itself. Without this fix, all of the time logged on issues
with a parent would have not been considered in the calculation.

On the other hand, including also the parent issue in the calculation does not hurt, since no time is ever logged
on a parent issue.