Reworked all queries
Hi,

this isn't compatible at all. It's just points out that there are many unneeded manual queries which were quite slow.

At all, the plugin isn't compatible with MS SQL Server. It would be great if you could take care of it. 

It seems that the issue statuses where broken on the index view. Maybe you could help with that?

Everything else seem to be okay.